Add test_jsonloggercontentloader::compareSessionLogConfigFileContEqual
Reason for all this refactoring is: It happened twice that we added a new
session but forgot to add logger content set information.

Once refactoring is done we will have per-session content files sharing the
same name as session file. Our initial test is just to compare the number of
session files with content set files.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>